Fixes: NB#172477 - Contactsd crash is observed while receiving IM messages during on-going voip call
Bug: 172477
RevBy: people-review
Details:

Two issues:
operator[] on QHash _inserts_ an empty element if the key isn't found, so
there's a chance to have a null pointer in that map. Replaced with
const_iterator and .find()

Second: There is a check for invalid pointer, but it is checked after the
first use.

The code of the original bug didn't have any check in there, so making sure
there's no null pointer in the list (at least accidently) and moving the check
should fix the bug.