changes: Fix remove IMAccount if it has no contact
The delete query for IMAccount was inside the loop for all contacts. That means
that if we have more than one contact, the query was repeated, and if we have no
contacts the query was not made at all.

Had to create the CDTpStorageRemoveAccount class to keep the AccountObjectPath for
the callback. Otherwise in the case we have no contact, it is impossible to know
the object path.