changes: Add CDTpContactPtr for shared ptr
If we get from telepathy an update on a contact, then the contact is removed,
that produced an error because the CDTpContact* was deleted while we resolve
the contact, so when onContactUpdateResolverFinished is called the CDTpContact*
was invalid. Using SharedPtr we make sure the object survive, and we set a flag on
it to know the contact was removed so we can skip the update.