Fixes: NB#257462 - Crash in literalIMAddress
When a contact or account is removed, cancel queued contact update if there were one.
The crash occurs because CDTpContact does not keep a ref to its CDTpAccount to avoid
a ref cycle, so if the queued contact update happens after the account is deleted,
we get a crash trying to access contact's account.