Fixes: Properly use deleteLater() in birthday controller
RevBy: cocos (MR#32)
Details: Only call deleteLater() when really not dealing with the object
anymore, otherwise the thread owning the object might delete it right after
we call deleteLater(), although we still deal with it in our callback. Actually
should not happen with this very code. Still makes sense to play by the rules.