Birthday plugin: Change testChangeBirthday to use a 3-day offset
If the test changes the birthday on the contact by only one day, it is
feasible that time zone differences could accumulate and make the test
fail in success conditions (or succeed in failure conditions). Change
the offset to be three days to make sure that this is not the case.