Revert "telepathy: Emit syncStarted even before the roster is received"
This reverts commit eca5505873645bef064f5c302dc262f8cd5894b9, I actually
have no idea why eca5505873645bef064f5c302dc262f8cd5894b9 made it to
master (maybe to send the syncStarted signal earlier since getting a
telepathy connection can take some time).

We can't emit the syncStarted signal as long as we don't get the
connection, since we need to test if this a roster-less account or not.