[contactsd-birthday-plugin] Improve birthday event handling
This commit ensures that the event summary for birthday events
is more descriptive, such as `Happy Birthday, John Smith!` for clarity.

It also removes any alarms set for birthday events, as birthdays are
commonly synced for a large number of friends from social services.

It also ensures that birthday events last for only one day.