[telepathy-plugin] Improve queued contact changeset binning
This commit ensures that if a queued contact change results in a new
contact being added, that the newly initialized contact is added to
the "All Changes" changeset bin of the contact save set.
This has the effect that any details generated during initialization
of the new contact (such as decomposed name details) will be stored
for the contact.