[contactsd] Add contacts DB exporter plugin
Create a plugin which listens for changes in the privileged contacts
database, and exports those changes to a nonprivileged database.

The nonprivileged database contains a simplified view of the contacts
on the device, which can be accessed by any application, since
restricted information is excluded from the export.

Changes made in the nonprivileged database are promoted to the
privileged database.