[contactsd] Process deletions before additions
A sync batch cannot include the same contact in both the deleted
set and either added or modified sets.  However, it is possible that
a contact could contain a detail, then be deleted, then a subsequent
creation/modification could include a detail that conflicts with the
first detail.  Therefore, we should process any deletions in the
batch prior to any other modifications, so that such cases do not
result in conflicts.