[contactsd] Only construct calendar and storage once. Contributes to JB#34931
contactsd includes a plugin which listens to changes to sync or
caldav accounts, and if such an account is disabled, it disables any
notebooks associated with that account.

The code which does this will construct a new calendar and storage
every time the change signal is received, which is at best inefficient
and at worst (e.g. if there is some resource leak within mkcal) lead
to eventual resource starvation (e.g. file descriptors, memory).