[contactsd] Fix errors when creating/removing telepathy account. Fixes JB#53425
Fixes two issues.

1. On account deletion the self contact is deleted with the account.
This triggered an error afterwards when updating the presence, during
which an attempt was being made to store the (no longer existing) self
contact, with the result that presence and accountList dbus signals were
not being emitted.

2. When creating a new account the self contact id wasn't being updated
because it was static, preventing any more than one account from being
properly created.