[contactsd] Replace the custom logging with qt tools. Fixes JB#53614
Qt logging is better what it used to be when all the existing code got
implemented. Nowadays better just use those.

Created a library for allowing the plugins to use that for common
plugin and logging code and moved the exported headers separate from
the rest. A little surprised that the symbol export from the main
executable even worked for that, and not sure did it even.

Didn't port the file logging, not aware of need and didn't spot
anything that would suggest using it. Can be added back based on need.